+++ android/src/main/java/com/marianhello/bgloc/data/ConfigurationDAO.java
@@ -3,9 +3,11 @@ package com.marianhello.bgloc.data;
 import java.util.Date;
 import java.util.Collection;
 
+import org.json.JSONException;
+
 import com.marianhello.bgloc.Config;
 
 public interface ConfigurationDAO {
-    public boolean persistConfiguration(Config config);
-    public Config retrieveConfiguration();
+    public boolean persistConfiguration(Config config) throws NullPointerException;
+    public Config retrieveConfiguration() throws JSONException;
 }