+++ ios/src/CDVBackgroundGeoLocation.m
@@ -351,9 +351,7 @@
     NSLog(@"- CDVBackgroundGeoLocation start (background? %ld)", (long)state);
     
     [locationManager startMonitoringSignificantLocationChanges];
-    if (state == UIApplicationStateBackground) {
-        [self setPace:isMoving];
-    }
+    [self setPace:isMoving];
     CDVPluginResult* result = nil;
     result = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
     [self.commandDelegate sendPluginResult:result callbackId:command.callbackId];